# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "SchoolBooks API",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://schoo-api.onrender.com/api",
            "description": "API Server"
        },
        {
            "url": "http://localhost:8081/api",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/saved-searches": {
            "get": {
                "tags": [
                    "SavedSearches"
                ],
                "summary": "Get user's saved searches",
                "description": "Returns current user's saved search queries",
                "operationId": "cebb1a91a448ea0d4a183f21481556ab",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/SavedSearch"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "SavedSearches"
                ],
                "summary": "Save a search query",
                "description": "Save a search query for the current user",
                "operationId": "153b18693c8e64c7ba01169c4c168c11",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "search_query"
                                ],
                                "properties": {
                                    "search_query": {
                                        "type": "string",
                                        "example": "javascript programming"
                                    },
                                    "search_name": {
                                        "type": "string",
                                        "example": "Programming Books",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Search saved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SavedSearch"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Search already saved or validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/saved-searches/{id}": {
            "get": {
                "tags": [
                    "SavedSearches"
                ],
                "summary": "Get specified saved search",
                "description": "Returns specified saved search (only if it belongs to current user)",
                "operationId": "13aab65b6ed07ce12b1d73c1e0d778e2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the saved search",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SavedSearch"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Saved search not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "SavedSearches"
                ],
                "summary": "Update saved search",
                "description": "Update a saved search name (only if it belongs to current user)",
                "operationId": "b853a3b4a883cd588ec1b0bdfd08a4e4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the saved search",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "search_name": {
                                        "type": "string",
                                        "example": "Updated Search Name"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Search updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SavedSearch"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Saved search not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "SavedSearches"
                ],
                "summary": "Delete saved search",
                "description": "Delete a saved search (only if it belongs to current user)",
                "operationId": "0e588be22905d069effe5b6501ed2624",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the saved search",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Search deleted successfully"
                    },
                    "404": {
                        "description": "Saved search not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/saved-searches/clear": {
            "delete": {
                "tags": [
                    "SavedSearches"
                ],
                "summary": "Clear all saved searches",
                "description": "Remove all saved searches for the current user",
                "operationId": "9f18838d4ec7f6c3f4bf4f389cd5a657",
                "responses": {
                    "200": {
                        "description": "All saved searches cleared successfully"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/saved-searches/count": {
            "get": {
                "tags": [
                    "SavedSearches"
                ],
                "summary": "Get saved searches count",
                "description": "Get the total number of saved searches for the current user",
                "operationId": "ed1b85e07b77af8ba13e3e3be6f527fb",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "count": {
                                            "type": "integer",
                                            "example": "5"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "SavedSearch": {
                "title": "SavedSearch",
                "description": "SavedSearch model",
                "properties": {
                    "id": {
                        "description": "Unique identifier for the saved search",
                        "type": "integer",
                        "example": "1"
                    },
                    "user_id": {
                        "description": "User ID associated with the saved search",
                        "type": "integer",
                        "example": "1"
                    },
                    "search_query": {
                        "description": "The search query string",
                        "type": "string",
                        "example": "javascript programming"
                    },
                    "search_name": {
                        "description": "Custom name for the saved search (optional)",
                        "type": "string",
                        "example": "Programming Books",
                        "nullable": true
                    },
                    "created_at": {
                        "description": "Date and time when the search was saved",
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-05-12 10:00:00"
                    },
                    "updated_at": {
                        "description": "Date and time when the search was last updated",
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-05-12 10:30:00"
                    },
                    "user": {
                        "$ref": "#/components/schemas/User"
                    }
                },
                "type": "object"
            },
...
            "User": {
                "title": "User",
                "description": "User model",
                "properties": {
                    "id": {
                        "description": "ID",
                        "type": "integer",
                        "example": "1"
                    },
                    "full_name": {
                        "description": "Full name",
                        "type": "string",
                        "example": "John Doe"
                    },
                    "email": {
                        "description": "Email",
                        "type": "string",
                        "format": "email",
                        "example": "johndoe@example.com"
                    },
                    "password": {
                        "description": "Password (hashed)",
                        "type": "string",
                        "format": "password",
                        "example": "$2y$10$..."
                    },
                    "phone_number": {
                        "description": "Phone number",
                        "type": "string",
                        "example": "+1234567890",
                        "nullable": true
                    },
                    "email_verified_at": {
                        "description": "Email verification timestamp",
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "created_at": {
                        "description": "Creation timestamp",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Last update timestamp",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}